% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_lm.R
\name{bootstrap_lm}
\alias{bootstrap_lm}
\title{Cross validation with a lm model}
\usage{
bootstrap_lm(df, family, n, formula, data_test)
}
\arguments{
\item{df}{Dataframe containing the outcome and features for training model}

\item{n}{number of cross validations to run}

\item{formula}{model formula used by caret}

\item{data_test}{test data set}

\item{method}{Model type used by caret}
}
\description{
Cross validation with a lm model
}
