% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_na_rate.R
\name{check_na_rate}
\alias{check_na_rate}
\title{Check the NA value rate for each column in a dataframe}
\usage{
check_na_rate(x)
}
\arguments{
\item{x}{a data frame}
}
\value{
a data frame with variable name and the NA rate
}
\description{
Check the NA value rate for each column in a dataframe
}
