% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi_sq_df.R
\name{chi_sq_test_df}
\alias{chi_sq_test_df}
\title{Apply chi-square test for any combination of two columns in a data set}
\usage{
chi_sq_test_df(data)
}
\arguments{
\item{data}{A dataframe whose columns will be tested for chi-square indepedence.}
}
\value{
a matrix with the p-values of the chi-square test. Each column or row represents a variable in the input dataframe
}
\description{
Apply chi-square test for any combination of two columns in a data set
}
\examples{
chi_sq_test_df(iris)
}
