% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_plot.R
\name{corr_plot}
\alias{corr_plot}
\title{Create a level plot to show correlations}
\usage{
corr_plot(data, increment, size, title = "")
}
\arguments{
\item{data}{Data frame object}

\item{increment}{increment for the legend label}

\item{size}{size of the axis labels}
}
\value{
the level plot
}
\description{
Create a level plot to show correlations
}
\examples{
corr_plot(iris,0.1,1)
}
