% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_caret.R
\name{cv_caret}
\alias{cv_caret}
\title{Cross validation with a classification model built by caret}
\usage{
cv_caret(df, method, n, formula, data_test, outcome_var)
}
\arguments{
\item{df}{Dataframe containing the outcome and features to train model}

\item{method}{Model type used by caret}

\item{n}{number of cross validations to run}

\item{formula}{a string to represent model formula used by caret}

\item{data_test}{test data set}
}
\description{
Cross validation with a classification model built by caret
}
\examples{
iris_select=iris\%>\%filter(Species \%in\% c('setosa','versicolor'))
iris_select$Species=as.character(iris_select$Species)
library(caret)
train_index=createDataPartition(iris_select$Species,times = 1,p = 0.8)$Resample1
data_test=iris_select[-train_index,]
data_train=iris_select[train_index,]
cv_caret(data_train,'rf',n=5,formula= 'Species~.',data_test=data_test,'Species')
}
