% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_lm.R
\name{cv_lm}
\alias{cv_lm}
\title{Cross validation with a lm model}
\usage{
cv_lm(df, family, n, formula, data_test, outcome_var)
}
\arguments{
\item{df}{Dataframe containing the outcome and features for training model}

\item{family}{Model type used by glm}

\item{n}{number of cross validations to run}

\item{formula}{model formula used by glm}

\item{data_test}{test data set}

\item{outcome_var}{a string to indicate the outcome variable name}
}
\description{
Cross validation with a lm model
}
\examples{
iris_select=iris\%>\%filter(Species \%in\% c('setosa','versicolor'))
iris_select$Species=as.character(iris_select$Species)
library(caret)
train_index=createDataPartition(iris_select$Species,times = 1,p = 0.8)$Resample1
data_test=iris_select[-train_index,]
data_train=iris_select[train_index,]
cv_lm(data_train,'binomial',n=5,formula= 'Species~.',data_test=data_test,'Species')
}
