% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_select.R
\name{feature_select}
\alias{feature_select}
\title{Select features with minimum threshold}
\usage{
feature_select(data_clust, min_count)
}
\arguments{
\item{data_clust}{a dataframe of features. No patient ID in this dataframe}

\item{min_count}{an integer}
}
\value{
a dataframe with ID, min_diag, patient_tenure
}
\description{
Select features with minimum threshold
}
