% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_numeric_2_binary.R
\name{normalize_numeric_2_binary}
\alias{normalize_numeric_2_binary}
\title{Normalize a continuous variable to a binary varibale based on median}
\usage{
normalize_numeric_2_binary(data, col_to_change, comparator)
}
\arguments{
\item{data}{a dataframe with some numeric columns}

\item{col_to_change}{column index to change from nuermic to binary}

\item{comparator}{> or >= median}
}
\value{
a dataframe with changed columns
}
\description{
Normalize a continuous variable to a binary varibale based on median
}
\examples{
normalize_numeric_2_binary(iris,c(1,2),'>')
}
